requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==4.2.16
django-cors-headers==4.5.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
